requirements.txt
numpy~=2.0.2
galois~=0.4.6
scipy~=1.13.1

pycryptodome~=3.23.0
